refactor(repository): ProductRepository를 인터페이스와 구현으로 분리
