+++ src/main/java/gift/service/ProductService.java
@@ -12,40 +12,37 @@ import org.springframework.stereotype.Service;
 @Service
 public class ProductService {
 
-    private final ProductRepository productRepository;
+    private final ProductRepository ProductRepository;
 
-    public ProductService(ProductRepository productRepository) {
-        this.productRepository = productRepository;
+    public ProductService(ProductRepository ProductRepository) {
+        this.ProductRepository = ProductRepository;
     }
 
     public List<ProductResponse> getAllProducts() {
-        return productRepository.findAll().stream()
+        return ProductRepository.findAll().stream()
                 .map(ProductResponse::new)
                 .collect(Collectors.toList());
     }
 
     public ProductResponse findProductById(Long id) {
-        return productRepository.findById(id)
+        return ProductRepository.findById(id)
                 .map(ProductResponse::new)
                 .orElseThrow(() -> new NoSuchElementException("해당 상품이 존재하지 않습니다."));
     }
 
     public ProductResponse addProduct(ProductRequest request) {
-        Product product = new Product(request.name(), request.price(),
-                request.imageUrl());
-        Product saved = productRepository.save(product);
+        Product product = new Product(request.name(), request.price(), request.imageUrl());
+        Product saved = ProductRepository.save(product);
         return new ProductResponse(saved);
     }
 
     public ProductResponse updateProduct(Long id, ProductRequest request) {
-        Product updated = productRepository.update(id, request.name(), request.price(),
+        Product updated = ProductRepository.update(id, request.name(), request.price(),
                 request.imageUrl());
         return updated != null ? new ProductResponse(updated) : null;
     }
 
     public boolean deleteProduct(Long id) {
-        return productRepository.delete(id);
+        return ProductRepository.delete(id);
     }
-
-
 }
